requirements.txt
python-nmap==0.7.1
netifaces==0.11.0
requests==2.31.0
ipaddress==1.0.23
tabulate==0.9.0
scapy==2.5.0
dnspython==2.4.2 